+++ libnautilus-extensions/nautilus-file-operations.c
@@ -622,7 +622,7 @@ extract_string_until (const char *original, const char *until_substring)
 {
 	char *result;
 	
-	g_assert (strlen (original) >= until_substring - original);
+	g_assert ((int) strlen (original) >= until_substring - original);
 	g_assert (until_substring - original >= 0);
 
 	result = g_malloc (until_substring - original + 1);
